Fix code

+++ examples/continuum.rs
@@ -102,7 +102,7 @@ fn main() {
 
     pointillism::create_from_sgn(
         "examples/continuum.wav",
-        NOTE_COUNT_LEN as u64 * note_len,
+        NOTE_COUNT_LEN * note_len,
         // 10.0 might be too much, but just to be safe from clipping.
         Volume::new(poly_loop, Vol::new(1.0 / 10.0)),
     )
